hospital_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
PyJWT==2.8.1
cryptography==41.0.7
django-filter==23.4
mysqlclient==2.2.0
